More fiddling with the sizes
